feat: Add hotel management functionality with serializer and UI integration
